refactor the claim module into sub-modules
