+++ workers/rust/container/src/lib.rs
@@ -1,8 +1,7 @@
 use bollard::Docker;
 use slog::{info, Logger};
-use std::marker::PhantomData;
 use taskcluster::Credentials;
-use taskcluster_lib_worker::claim::{WorkClaimer, WorkClaimerConfig};
+use taskcluster_lib_worker::claim::WorkClaimer;
 use taskcluster_lib_worker::process::ProcessFactory;
 
 mod executor;
@@ -19,17 +18,15 @@ pub async fn main(logger: Logger) -> anyhow::Result<()> {
     let docker = Docker::connect_with_local_defaults()?;
     let executor = ContainerExecutor::new(docker);
 
-    let wc = WorkClaimer::new(WorkClaimerConfig {
-        logger: logger.clone(),
-        root_url: root_url.to_owned(),
-        worker_creds: Credentials::from_env()?,
-        task_queue_id: "aa/bb".to_owned(),
-        worker_group: "rust".to_owned(),
-        worker_id: "worker".to_owned(),
-        capacity: 4,
-        executor,
-        payload_type: PhantomData,
-    });
+    let wc = WorkClaimer::new(executor)
+        .logger(logger.clone())
+        .root_url(root_url.to_owned())
+        .worker_creds(Credentials::from_env()?)
+        .task_queue_id("aa/bb".to_owned())
+        .worker_group("rust".to_owned())
+        .worker_id("worker".to_owned())
+        .capacity(4)
+        .build();
     let wc = wc.start();
     wc.await
 }
