+++ workers/rust/lib-worker/src/claim/mod.rs
@@ -1,0 +1,6 @@
+mod long_poll;
+mod task_claim;
+mod work_claimer;
+
+pub(crate) use task_claim::TaskClaim;
+pub use work_claimer::{WorkClaimer, WorkClaimerBuilder};
