+++ workers/rust/lib-worker/src/lib.rs
@@ -14,9 +14,7 @@ pub mod executor;
 pub mod process;
 pub mod task;
 
-// TODO: basic docker support
 // TODO: worker-runner protocol
 // TODO: artifact support
 // TODO: live-logging support (websocktunnel)
-// TODO: move more of executor into lib-worker, to handle creds rotation, posting status, artifact upload, etc.
 // TODO: testing utilities
